requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file development.txt development.in
#
click==6.7                # via pip-tools
defusedxml==0.4.1         # via python3-openid
Django==1.10.5
djangorestframework==3.5.3
first==2.0.1              # via pip-tools
flake8==3.2.1
mccabe==0.5.3             # via flake8
oauthlib==2.0.1           # via requests-oauthlib, social-auth-core
pip-tools==1.8.0
pycodestyle==2.2.0        # via flake8
pyflakes==1.3.0           # via flake8
PyJWT==1.4.2              # via social-auth-core
python-social-auth==0.3.3
python3-openid==3.0.10    # via social-auth-core
requests-oauthlib==0.7.0  # via social-auth-core
requests==2.13.0          # via requests-oauthlib, social-auth-core
six==1.10.0               # via pip-tools, social-auth-app-django, social-auth-core
social-auth-app-django==0.1.0
social-auth-core==1.0.1
